@@ -1,3 +1,3 @@
 module shanhu.io/gcimporter
 
-go 1.17
+go 1.18
